Move DebugOptions to Storage namespace
