+++ visitz/Visitz/Services/Authentication/Keycloak/VisitzSessionInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
+using Visitz.Storage;
 
 namespace Visitz.Services.Authentication.Keycloak
 {
